turkish_pill/requirements.txt
# PyTorch (Nightly - RTX 5060 uyumu için)
--extra-index-url https://download.pytorch.org/whl/nightly/cu124
torch>=2.4.0
torchvision>=0.19.0

# Transformers (ViT için)
transformers>=4.40.0
accelerate>=0.30.0

# Görüntü İşleme
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Streamlit
streamlit>=1.32.0

# Yardımcı Kütüphaneler
pyyaml>=6.0
tqdm>=4.66.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Data augmentation
albumentations>=1.4.0

